api/requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
Werkzeug==2.0.1
pandas
tqdm
tritonclient[grpc]
pillow
